+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Upgrade the `org.realityforge.org.jetbrains.annotations` artifact to version `1.2.0`.
 * Upgrade the `org.realityforge.react4j` artifacts to version `0.139`.
 * Remove some unused methods from `EntityMessageCacheUtil`.
+* Remove unused method `ReplicantContextHolder.contains(...)`.
 
 ### [v6.48](https://github.com/replicant4j/replicant/tree/v6.48) (2019-10-18) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.47...v6.48)